requirements.txt
Django==2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.8.0
Pillow==5.0.0
requests==2.18.4